+++ meal-order/src/main/java/com/alfredamos/meal_order/controllers/UserController.java
@@ -2,7 +2,7 @@ package com.alfredamos.meal_order.controllers;
 
 
 import com.alfredamos.meal_order.dto.UserDto;
-import com.alfredamos.meal_order.mapper.UserMapper;
+import com.alfredamos.meal_order.exceptions.ForbiddenException;
 import com.alfredamos.meal_order.services.UserService;
 import com.alfredamos.meal_order.utils.ResponseMessage;
 import io.swagger.v3.oas.annotations.tags.Tag;
@@ -19,25 +19,17 @@ import java.util.UUID;
 @RestController
 @RequestMapping("/api/users")
 public class UserController {
-
     private final UserService userService;
+    private final OwnerCheck ownerCheck;
 
-    private final UserMapper userMapperImpl;
 
-    @GetMapping
-    public ResponseEntity<List<UserDto>> getAllUsers(){
-       var users = this.userService.getAllUsers();
-
-        var usersDto = this.userMapperImpl.toDTOList(users);
-
-       return new ResponseEntity<>(usersDto, HttpStatus.OK);
-    }
-    
     @GetMapping("/{id}")
     public ResponseEntity<UserDto> getUserById(@PathVariable UUID id){
-        var user = this.userService.getUserById(id).orElse(null);
-
-        var userDto = this.userMapperImpl.toDTO(user);
+        var isSameUser = ownerCheck.compareUserId(id);
+        if (!isSameUser){
+            throw new ForbiddenException("You are not permitted to view this resource!");
+        }
+        var userDto = this.userService.getUserById(id);
 
         return new ResponseEntity<>(userDto, HttpStatus.OK);
         
@@ -45,6 +37,11 @@ public class UserController {
 
     @DeleteMapping("/{id}")
     public ResponseEntity<ResponseMessage> deleteUserById(@PathVariable UUID id){
+        var isSameUser = ownerCheck.compareUserId(id);
+        if (!isSameUser){
+            throw new ForbiddenException("You are not permitted to delete this resource!");
+        }
+
         var responseMessage = this.userService.deleteUser(id);
 
         return new ResponseEntity<>(responseMessage, HttpStatus.OK);
